bazel: handle runfiles init with global constructor
No need to modify OpenROAD or OpenSTA source code
