+++ bazel/BUILD
@@ -6,3 +6,14 @@ compile_pip_requirements(
     python_version = "3.13",
     requirements_txt = "requirements_lock_3_13.txt",
 )
+
+cc_library(
+    name = "runfiles",
+    srcs = [
+        "InitRunFiles.cpp",
+    ],
+    deps = [
+        "@rules_cc//cc/runfiles",
+    ],
+    visibility = ["//visibility:public"],
+)
