+++ src/Main.cc
@@ -29,10 +29,6 @@
 #include <tclExtend.h>
 #endif
 
-#ifdef BAZEL_CURRENT_REPOSITORY
-#include "rules_cc/cc/runfiles/runfiles.h"
-#endif
-
 #include "gui/gui.h"
 #include "ord/Design.h"
 #include "ord/InitOpenRoad.hh"
@@ -198,33 +194,6 @@ static void handler(int sig)
   raise(sig);
 }
 
-#ifdef BAZEL_CURRENT_REPOSITORY
-
-// Avoid adding any dependencies like boost.filesystem
-//
-// Returns path to running binary if possible, otherwise nullopt.
-static std::optional<std::string> getProgramLocation()
-{
-#if defined(_WIN32)
-  char result[MAX_PATH + 1] = {'\0'};
-  auto path_len = GetModuleFileNameA(NULL, result, MAX_PATH);
-#elif defined(__APPLE__)
-  char result[MAXPATHLEN + 1] = {'\0'};
-  uint32_t path_len = MAXPATHLEN;
-  if (_NSGetExecutablePath(result, &path_len) != 0) {
-    path_len = readlink("/proc/self/exe", result, MAXPATHLEN);
-  }
-#else
-  char result[PATH_MAX + 1] = {'\0'};
-  ssize_t path_len = readlink("/proc/self/exe", result, PATH_MAX);
-#endif
-  if (path_len > 0) {
-    return result;
-  }
-  return std::nullopt;
-}
-#endif
-
 int main(int argc, char* argv[])
 {
   // This avoids problems with locale setting dependent
@@ -236,19 +205,6 @@ int main(int argc, char* argv[])
       break;
     }
   }
-
-#ifdef BAZEL_CURRENT_REPOSITORY
-  using rules_cc::cc::runfiles::Runfiles;
-  std::string error;
-  std::unique_ptr<Runfiles> runfiles(Runfiles::Create(
-      getProgramLocation().value(), BAZEL_CURRENT_REPOSITORY, &error));
-  if (!runfiles) {
-    std::cerr << error << std::endl;
-    return 1;
-  }
-  std::string path = runfiles->Rlocation("tk_tcl/library/");
-  setenv("TCL_LIBRARY", path.c_str(), 0);
-#endif
 
   // Generate a stacktrace on crash
   signal(SIGABRT, handler);
